[nrf noup] Add tfm read service to nRF SoC Flash driver
This adds the functionality to read memory which are residing in the
secure part of flash from a non-secure application when building with
TF-M.

Currently we check if the address space is not inside the non-secure
applapplication address space. If the address space is outside we do a
request to TF-M so that the secure application can either return success
or fail if the address space is not allowed to be read.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>