[nrf fromlist] bluetooth: gatt: fix leaking buffers on send error
bt_att_send() releases a buffer only in some error
scenarios, so users of the function are unable to make
a right decision whether to release the buffer or not.
This might lead to buffer leaks and deadlocks during DFU
over SMP/BLE (in fact, it has been observed).

Make the behaviour consistent, that is, let the caller to
retain the ownership of a buffer when bt_att_send() returns
an error. Then fix all the function users to release
a buffer properly.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>
(cherry picked from commit dd911fb2eda3ea1a5e05da4e96bdb672170d4101)