[nrf noup] ci: clang: parallel execution
Bring back parallel execution to the clang action. In the upstream
they went to use a single dedicated powerfull agent to run this. We
need to keep this as we have to base on agents available in github's
cloud.

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>
Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>
Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>
(cherry picked from commit 914d890a5e871e234a7093fa31b65e694c516559)