[nrf fromtree] Bluetooth: hci: correct ext adv cmd definition
Update the definition of the set extended advertising data command to be
a variable array instead of hardcoded to the maximum length. This
conforms to the definition from the Bluetooth specification and allows
the corresponding code to be slightly cleaner.

Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
(cherry picked from commit 855f4b9c963f844a10c81c65aefe3bd763c2f0d3)
Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>