[nrf fromtree] boards: arduino_nano_33_ble: Correct assignment of P0.13 to PWM
This is a follow-up to commit f4a0ddd8afa1c2b99a30dcd940f34304ce678c93.

Since the yellow LED and the SCK line of spi2 use the same pin
(P0.13), they cannot be used together. Consequently, the pin
should not be assigned to the same PWM instance as other pins
that drive LEDs, as the limitation of usage would apply to the
whole PWM instance (it acquires all the pins assigned to it on
initialization of the PWM driver, regardless of whether the PWM
signal is eventually generated on particular outputs or not).

Use a separate PWM instance (disabled by default) for driving
the yellow LED. Don't use the "nordic,invert" property for that
PWM, as the yellow LED is active high.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
(cherry picked from commit 0cf57b129b1d545746e83aa992b65fe5ecf7f03a)