[nrf fromlist] bluetooth: host: Fix reading current PHY when using extended advertising
The current PHY is read only when automatic PHY
update procedure is enabled, 2M PHY is supported and
extended advertising is enabled and this is done after
application is notified about connection. This leads
to invalid connection info PHY data which always returns
1M PHY because when using extended advertising the connection
can be established on different PHY and host does not read
this value from the controller in most cases. In order to
have a current PHY value updated we need to read it in
connection complete event before the user application
is notified about connection.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/46113

Signed-off-by: Kamil Gawor <Kamil.Gawor@nordicsemi.no>