[nrf fromlist] tfm: Add option for implementation custom reset handler with TF-M
Zephyr adds a custom handler that overrides the weak function
sys_arch_reset when TF-M platform partition is enabled.

This takes away the option for the application to override the weak
definition for their platform or use-case.

Add an option that control whether this default reset handling is added
to the build.

Upstream PR:
https://github.com/zephyrproject-rtos/zephyr/pull/46133

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>