subsys: trusted_storage: add Internal Trusted Storage API with ChaChaPoly Trust
This adds the Internal Trusted Storage API which is designed to
store sentitive data like keys for use by components of the PSA Root of Trust.

The Internal Trusted Storage uses the same storage backend as Protected
Storage with a different prefix to not overlap the Objects IDs.

The Internal Trusted Storage has a strong requirement of trust, so only
the Authenticated Encryption with AEAD scheme is implemented.

This API is not designed to be called directly by the application but
by a PSA Crypto implementation like MbedTLS to store the non-volatile
keys.

Since it's designed to be called from within a PSA Crypto implementation,
using PSA Crypto API as AEAD implementation is not recommended so the
initial implementation uses the legacy MbedTLS API, but the configuration
is made easy to implement support for platform specific AEAD APIs.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>