Bluetooth: Audio: Fix PACS UUID issue
We attempted to simply store the UUID in a local variable
before sending it to bt_gatt_notify_uuid, but evidently
the `bt_uuid` struct does not work that way, and it
ended up giving errors.

Fixed by providing the BT_UUID_PACS_X UUIDs directly as
arguments, instead of storing in a local variable first.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>