boards: Configure I2C/SPI default for boards with magnetometer alias
For all boards that have a magnetometer devicetree alias, configure the
appropriate bus (I2C or SPI) default to be enabled when the sensor
driver class is enabled. This will simplify enabling the magn_polling
sample application for these boards (and using the magnetometer in
general) because it will eliminate the need to add a bunch of
board-specific configuration overlays to the application.

Signed-off-by: Maureen Helm <maureen.helm@intel.com>
(cherry picked from commit adc37bad40dc5538b31343890bd1aad681b0d1f3)