test/arch: Remove xtensa_asm2 test
This test has been bitrotting for a while - years, in fact. As it didn't
have a `testcase.yaml`, it wasn't being run on twister. The code expects
to be run on single thread environment, but xtensa arch currently
doesn't support that.

This patch simply removes this test - if it's deemed important, it can
be reworked and readded in the future.

Fixes #47508.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>
(cherry picked from commit 63a497280f22aea39793dd9de9731469a83ce533)