tests: logging: disable logging stack test
This test keeps breaking whenever something in kernel or
architecture code changes. If we are to track footprint and stack, it
should be done in some other way.
The value of the test deminished if we have to change the boundaries
every second day.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>
(cherry picked from commit d0be1de99e8e95df91f9d4a8f97e6474bce11230)