[nrf fromtree] boards: Configure I2C/SPI default for boards with accelerometer
For boards with an accelerometer, configure the appropriate bus
(I2C or SPI) default to be enabled if the sensor driver class was enabled.

This change makes no need to add a bunch of board-specific configuration
overlays to the application.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
(cherry picked from commit 342f74a7923acd14123abde9f433e71edc09f4bf)