[nrf fromtree] Bluetooth: Host: Get first packet matching channel
When sending an ATT packet, do not just try to send the first packet in
the queue, but find the first one that can be sent using the channel.
This makes sure that eg. packets that shall be sent on an unenhanced ATT
channel are not blocked by packets that shall be sent on enhanced ATT
channels.

Signed-off-by: Herman Berget <herman.berget@nordicsemi.no>
(cherry picked from commit 62ff17df2134d4faa9a9600f1ee634b7ebef47aa)
Signed-off-by: Herman Berget <herman.berget@nordicsemi.no>