[nrf fromtree] boards: nrf: Activate pull-up on QSPI CSN line in "sleep" state
When there is no external pull-up on the CSN line and the line
is put into the low-power mode (input direction and disconnected
buffer), after some time its level falls to low and some flash
chips, like for example MX25R64, exit the Deep Power-down mode
when that happens. To prevent this, activate the GPIO internal
pull-up for this line in boards that do not have the external
pull-up on it.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
(cherry picked from commit 65976137dcde4536a53ffe3c4a92a2e7debdd66a)