[nrf fromlist] drivers: sensor: adxl372: Fix default bus issue
The adxl372 device does not select the correct bus that it is on, which
causes undefined operation (usually a fault). This driver should not be
using Kconfig to determine which bus to use, that should be done by
dts, however this works as a quick fix in the interim until the driver
can be fixed properly.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/49018

Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>